requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
bidict==0.23.1
Booktype==1.5
BTrees==6.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
Django==4.2.15
h11==0.14.0
idna==3.8
persistent==6.0
pycparser==2.22
pyscard==2.0.10
python-engineio==4.9.1
python-socketio==5.11.3
redis==5.0.8
requests==2.32.3
simple-websocket==1.0.0
simplejson==3.19.3
smartcard==0.3
sqlparse==0.5.1
transaction==4.0
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
zc.lockfile==3.0.post1
ZConfig==4.1
ZODB==6.0
zodbpickle==4.0
zope.deferredimport==5.0
zope.interface==7.0.3
zope.proxy==5.3
